requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
